+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.36 fix: fix option_finance_board interface
+
+    1. 修复 option_finance_board 接口
+
 1.6.35 fix: fix stock_zh_index_hist_csindex interface
 
     1. 修复 stock_zh_index_hist_csindex 接口
@@ -829,6 +833,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.36 fix: fix option_finance_board interface
 
 1.6.35 fix: fix stock_zh_index_hist_csindex interface